RK3399 SMP bring-up, docs, tooling, CI; stop tracking build artifacts
- SMP/CCI/GIC work, flash/recovery scripts, TF-A skip-M0 patch note
- third_party/ gitignored (~GB); see patches/THIRD_PARTY.txt
- GitHub Actions: build kernel_neuro with aarch64-linux-gnu
- Inline _nanopi_boot_src tree (remove broken submodule gitlink)

+++ hal/uart.c
@@ -32,7 +32,7 @@ static inline u32 uart_read_reg(volatile u32* base, u32 offset) {
 result_t uart_init(uart_t* uart, const uart_config_t* config) {
     if (!uart || !config) return ERR_INVALID_PARAM;
     
-    uart->base = (volatile u32*)config->base_addr;
+    uart->base = (volatile u32*)(uintptr_t)config->base_addr;
     uart->config = *config;
     
     // Calculate divisor for baud rate (24MHz clock)
